# perf: Replace `tibble()` by `fast_tibble()`