--- a/R/utils.R
+++ b/R/utils.R
@@ -32,6 +32,11 @@ tail.dm_zoomed <- function(x, n = 6L, ...) {
   replace_zoomed_tbl(x, tail(tbl_zoomed(x), n, ...))
 }
 
+fast_tibble <- function(...) {
+  out <- vctrs::df_list(...)
+  tibble::new_tibble(out)
+}
+
 recode_compat <- function(x, y) {
   x <- unname(x)
   idx <- vctrs::vec_match(names(y), x)