There are other places where `tibble()` is used. It might be worth looking into them as well.

```r
data <- nycflights_subset()
flights <- data$flights
weather <- data$weather
airlines <- data$airlines
airports <- data$airports
planes <- data$planes

dm <- dm(airlines, airports, flights, planes, weather)

add_keys <- function() {
  dm %>% 
    dm_add_pk(planes, tailnum) %>%
    dm_add_pk(airlines, carrier) %>%
    dm_add_pk(airports, faa) %>%
    dm_add_fk(flights, tailnum, planes) %>%
    dm_add_fk(flights, carrier, airlines) %>%
    dm_add_fk(flights, origin, airports)
}

for (i in 1:100) {
  dm_nycflights13()
}

bench::mark(
  add_keys()
)

# before
# # A tibble: 1 × 13
#   expression     min median `itr/sec` mem_alloc `gc/sec` n_itr
#   <bch:expr> <bch:t> <bch:>     <dbl> <bch:byt>    <dbl> <int>
# 1 add_keys()  9.69ms 10.4ms      90.1    35.1KB     13.7    33

# after
# # A tibble: 1 × 13
#   expression     min median `itr/sec` mem_alloc `gc/sec` n_itr
#   <bch:expr> <bch:t> <bch:>     <dbl> <bch:byt>    <dbl> <int>
# 1 add_keys()  5.31ms 7.31ms      133.     700KB     8.89    60
```